pyproject.toml
[project]
name = "prep"
version = "0.1.0"
description = "Primed"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "supabase>=2.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "qstash>=3.2.0",
    "trycourier>=6.2.1",
    "posthog>=6.7.6",
    "python-multipart>=0.0.20",
    "google-adk>=0.5.0",
    "google-genai>=1.41.0",
    "google-generativeai>=0.8.5",
    "websockets>=14.0",
    "python-jose[cryptography]>=3.5.0",
    "slowapi>=0.1.9",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.2",
]
